requirements.txt
# Core Dependencies
python-telegram-bot==20.3  # Uses httpx 0.24.x
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1

# XRP Ledger
xrpl-py==2.5.0  # Compatible with httpx <0.25.0

# Security
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9

# API and HTTP
httpx==0.24.1  # Works with both PTB 20.3 and xrpl-py 2.5.0
requests==2.31.0

# Rate Limiting
slowapi==0.1.9

# Caching
redis==5.0.1

# Monitoring and Error Tracking
sentry-sdk[fastapi]==2.19.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==19.13.0
pandas==2.2.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
types-sqlalchemy==1.4.53.38

# Logging
python-json-logger==2.0.7